ai_backend/requirements.txt
azure-data-tables==12.6.0
azure-storage-blob==12.24.1
fastapi==0.115.6
langchain==0.3.14
langchain-community==0.3.14
langchain-openai==0.3.0
openai==1.59.7
openpyxl
pandas==2.2.3
pdfplumber
pydantic==2.10.3
python-docx
python-dotenv==1.0.1
python-multipart
python-pptx
requests==2.32.3
uvicorn==0.34.0
websockets==14.1
azure-cosmos==4.9.0